move grouping info helpers to separate module
